🛠️ Chat과 Message는 데이터 클래스로 추출
